requirements.txt
certifi==2021.5.30
chardet==4.0.0
ciso8601==2.1.3
idna==2.10
python-dotenv==0.17.1
requests==2.25.1
urllib3==1.26.5
